refactor(pipeline/pii): split Personal-data redactors (Phone/Name/Merchant) into per-category modules
Phase 6 commit 3/7 of PiiRedactor split.

- Carve phone/name/merchant from src/Scrapers/Pipeline/Types/PiiRedactor.ts.
- New files: src/Scrapers/Pipeline/Types/PiiRedactor/{Phone,Name,Merchant}.ts.
- New shared helper: src/Scrapers/Pipeline/Types/PiiRedactor/CommonHelpers.ts
  hosts graphemeCount + buildSegmenter (used by Name + Merchant +
  redactErrorMessage + replaceValueAttr).
- PiiRedactor.ts re-exports redactPhone/Name/Merchant from the new
  modules and imports graphemeCount from CommonHelpers for its still-
  resident error-log + HTML helpers.
- Tests: zero body diff — 138 PiiRedactor tests pass unchanged.

Verified:
- npm run build green
- npm run test:pipeline (PiiRedactor cluster) green
- npm run lint + npm run type-check green
- Each new file <= 80 LoC

Master plan: pipeline-decoupling-master-2026-05-28 / phase-6
